test: 복잡한 테스트 픽스처 LinesFixture에서 생성
