+++ src/main/java/wooteco/subway/domain/Line.java
@@ -2,6 +2,7 @@ package wooteco.subway.domain;
 
 import java.util.LinkedList;
 import java.util.List;
+import java.util.Objects;
 
 public class Line {
 
@@ -76,5 +77,22 @@ public class Line {
 
 	public List<Section> getSections() {
 		return sections.getValues();
+	}
+
+	@Override
+	public boolean equals(Object o) {
+		if (this == o)
+			return true;
+		if (o == null || getClass() != o.getClass())
+			return false;
+		Line line = (Line)o;
+		return Objects.equals(getId(), line.getId()) && Objects.equals(getName(), line.getName())
+			&& Objects.equals(getColor(), line.getColor()) && Objects.equals(getSections(),
+			line.getSections());
+	}
+
+	@Override
+	public int hashCode() {
+		return Objects.hash(getId(), getName(), getColor(), getSections());
 	}
 }
