+++ src/test/java/wooteco/subway/repository/SectionsDirtyCheckerTest.java
@@ -1,4 +1,4 @@
-package wooteco.subway.domain;
+package wooteco.subway.repository;
 
 import static org.assertj.core.api.Assertions.*;
 import static org.junit.jupiter.api.Assertions.*;
@@ -8,6 +8,9 @@ import java.util.List;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 
+import wooteco.subway.domain.Section;
+import wooteco.subway.domain.Sections;
+import wooteco.subway.domain.Station;
 import wooteco.subway.repository.SectionsDirtyChecker;
 
 class SectionsDirtyCheckerTest {
