+++ src/test/java/wooteco/subway/service/PathServiceTest.java
@@ -1,10 +1,7 @@
 package wooteco.subway.service;
 
-import static java.util.stream.Collectors.*;
 import static org.assertj.core.api.Assertions.*;
 
-import java.util.Arrays;
-import java.util.HashMap;
 import java.util.List;
 import java.util.Map;
 
@@ -15,9 +12,8 @@ import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.test.context.SpringBootTest;
 import org.springframework.transaction.annotation.Transactional;
 
-import wooteco.subway.domain.Line;
+import wooteco.subway.LinesFixture;
 import wooteco.subway.domain.Path;
-import wooteco.subway.domain.Section;
 import wooteco.subway.domain.Station;
 
 @SpringBootTest
@@ -25,12 +21,10 @@ import wooteco.subway.domain.Station;
 class PathServiceTest {
 
 	@Autowired
-	private StationService stationService;
-	@Autowired
-	private LineService lineService;
-	@Autowired
 	private PathService pathService;
-	private final Map<String, Station> stations = new HashMap<>();
+	@Autowired
+	private LinesFixture linesFixture;
+	private Map<String, Station> stations;
 
 	//            신사 (3호선)
 	//            |10|
@@ -42,40 +36,7 @@ class PathServiceTest {
 
 	@BeforeEach
 	void init() {
-		stations.putAll(saveStations("내방역", "고속터미널역", "반포역", "논현역"));
-
-		Line line7 = lineService.create("7호선", "red",
-			new Section(stations.get("내방역"), stations.get("고속터미널역"), 10));
-		lineService.addSection(line7.getId(),
-			makeSection(stations.get("고속터미널역"), stations.get("반포역"), 10));
-		lineService.addSection(line7.getId(),
-			makeSection(stations.get("반포역"), stations.get("논현역"), 10));
-
-		stations.putAll(saveStations("신사역", "잠원역", "서초역"));
-
-		Line line3 = lineService.create("3호선", "red",
-			new Section(stations.get("신사역"), stations.get("잠원역"), 10));
-		lineService.addSection(line3.getId(),
-			makeSection(stations.get("잠원역"), stations.get("고속터미널역"), 10));
-		lineService.addSection(line3.getId(),
-			makeSection(stations.get("고속터미널역"), stations.get("서초역"), 10));
-
-		Station sapyung = stationService.create("사평역");
-		stations.put("사평역", sapyung);
-		lineService.create("9호선", "red", makeSection(
-			stations.get("고속터미널역"), sapyung, 14));
-
-		lineService.create("새호선", "red",
-			makeSection(stations.get("서초역"), sapyung, 3));
-	}
-
-	private Map<String, Station> saveStations(String... names) {
-		return Arrays.stream(names)
-			.collect(toMap(name -> name, name -> stationService.create(name)));
-	}
-
-	private Section makeSection(Station source, Station target, int distance) {
-		return new Section(source, target, distance);
+		stations = linesFixture.initAndReturnStationMap();
 	}
 
 	@DisplayName("한 라인에서 조회한다.")
